% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milpplot.R
\name{milp_plot}
\alias{milp_plot}
\title{Plot a mixed-integer linear program}
\usage{
milp_plot(A, b, cv)
}
\arguments{
\item{A}{the constraint matrix or something that can be coerced to a matrix (e.g. a sparse matrix).}

\item{b}{the right-hand side vector b.}

\item{cv}{the objective vector c.}
}
\value{
a ggplot2 object
}
\description{
Plot a mixed-integer linear program
}
